+++ apollo-accounts/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'nicolaslopezj:apollo-accounts',
-  version: '1.3.0',
+  version: '1.3.1',
   // Brief, one-line summary of the package.
   summary: 'Meteor accounts in GraphQL',
   // URL to the Git repository containing the source code for this package.
@@ -17,10 +17,9 @@ Package.onUse(function (api) {
   api.use('npm-bcrypt')
   api.use('ecmascript')
   api.use('random')
-  api.use('orionsoft:graphql-compiler@0.0.1')
 
   // Oauth
-  api.use(['http', 'oauth', 'facebook', 'accounts-facebook'], {weak: true})
+  api.use(['http', 'oauth', 'facebook', 'accounts-facebook', 'google', 'accounts-google'], {weak: true})
 
   api.mainModule('src/index.js')
 })